created new file table.js
